Move take course button frontend code
