+++ assets/blocks/single-course.scss
@@ -1,3 +1,3 @@
 
-@import 'button/button.scss';
-@import 'course-outline/style.scss';
+@import 'take-course/take-course';
+@import 'course-outline/style';
