+++ assets/blocks/take-course/take-course.scss
@@ -1,7 +1,3 @@
-
-.wp-block-sensei-button {
-
-}
 .wp-block-sensei-lms-button-take-course {
 
 	.block-editor-rich-text {
